control-service: add IT tests for async job deployment (#2794)
I've introduced the initial version of async job deployment integration
tests. This is just the starting point, and we plan to expand and
refactor it further in the upcoming version based on feedback from
@antoniivanov 's comment. You can find the discussion on this change in
the following link: [Link to the GitHub pull request
discussion](https://github.com/vmware/versatile-data-kit/pull/2781#discussion_r1358096950).


---------
