+++ projects/control-service/projects/pipelines_control_service/src/main/java/com/vmware/taurus/datajobs/DeploymentModelConverter.java
@@ -208,7 +208,7 @@ public class DeploymentModelConverter {
   }
 
   public static DesiredDataJobDeployment mergeDeployments(
-      ActualDataJobDeployment oldDeployment, JobDeployment newDeployment, String userDeployer) {
+      DesiredDataJobDeployment oldDeployment, JobDeployment newDeployment, String userDeployer) {
     checkDeploymentsCanBeMerged(oldDeployment, newDeployment);
     DesiredDataJobDeployment mergedDeployment = new DesiredDataJobDeployment();
     mergedDeployment.setDataJobName(newDeployment.getDataJobName());
@@ -240,7 +240,7 @@ public class DeploymentModelConverter {
   }
 
   private static void checkDeploymentsCanBeMerged(
-      ActualDataJobDeployment oldDeployment, JobDeployment newDeployment) {
+      DesiredDataJobDeployment oldDeployment, JobDeployment newDeployment) {
     if (oldDeployment.getDataJobName() == null
         || newDeployment.getDataJobName() == null
         || newDeployment.getDataJobTeam() == null
@@ -263,7 +263,7 @@ public class DeploymentModelConverter {
   private static void mergeDeploymentResources(
       DesiredDataJobDeployment mergedDeployment,
       JobDeployment newDeployment,
-      ActualDataJobDeployment oldDeployment) {
+      DesiredDataJobDeployment oldDeployment) {
     if (newDeployment.getResources() == null) {
       return;
     }
