+++ projects/control-service/projects/pipelines_control_service/src/main/java/com/vmware/taurus/service/deploy/DeploymentServiceV2.java
@@ -16,6 +16,7 @@ import com.vmware.taurus.service.model.DataJob;
 import com.vmware.taurus.service.model.DeploymentStatus;
 import com.vmware.taurus.service.model.DesiredDataJobDeployment;
 import com.vmware.taurus.service.model.JobDeployment;
+import com.vmware.taurus.service.model.JobConfig;
 import com.vmware.taurus.service.notification.NotificationContent;
 import com.vmware.taurus.service.repository.ActualJobDeploymentRepository;
 import com.vmware.taurus.service.repository.DesiredJobDeploymentRepository;
@@ -63,7 +64,8 @@ public class DeploymentServiceV2 {
    */
   public void patchDesiredDbDeployment(
       DataJob dataJob, JobDeployment jobDeployment, String userDeployer) {
-    actualJobDeploymentRepository
+    populateDeploymentForBackwardCompatibility(dataJob, jobDeployment);
+    desiredJobDeploymentRepository
         .findById(dataJob.getName())
         .ifPresentOrElse(
             oldDeployment -> {
@@ -87,7 +89,8 @@ public class DeploymentServiceV2 {
    */
   public void updateDesiredDbDeployment(
       DataJob dataJob, JobDeployment jobDeployment, String userDeployer) {
-    actualJobDeploymentRepository
+    populateDeploymentForBackwardCompatibility(dataJob, jobDeployment);
+    desiredJobDeploymentRepository
         .findById(dataJob.getName())
         .ifPresentOrElse(
             oldDeployment -> patchDesiredDbDeployment(dataJob, jobDeployment, userDeployer),
@@ -258,5 +261,43 @@ public class DeploymentServiceV2 {
   private boolean deploymentExistsOrInProgress(String dataJobName) {
     return jobImageBuilder.isBuildingJobInProgress(dataJobName)
         || readDeployment(dataJobName).isPresent();
+  }
+
+  /**
+   * Populates the deployment's fields with legacy data to ensure backward compatibility. This
+   * method is used to maintain compatibility with older versions of the deployment process.
+   */
+  private void populateDeploymentForBackwardCompatibility(
+      DataJob dataJob, JobDeployment jobDeployment) {
+    setDeploymentSchedule(dataJob, jobDeployment);
+    setDeploymentEnabled(jobDeployment);
+  }
+
+  /**
+   * Sets the deployment schedule if it is not passed to the API for backward compatibility, as the
+   * schedule is still passed as part of the data job.
+   *
+   * @param dataJob the data job to which the deployment is associated
+   * @param jobDeployment the deployment to patch with
+   */
+  private void setDeploymentSchedule(DataJob dataJob, JobDeployment jobDeployment) {
+    if (jobDeployment.getSchedule() == null) {
+      String schedule =
+          Optional.ofNullable(dataJob)
+              .map(DataJob::getJobConfig)
+              .map(JobConfig::getSchedule)
+              .orElse(null);
+      jobDeployment.setSchedule(schedule);
+    }
+  }
+
+  /**
+   * Sets the deployment enabled flag to true if it is not passed to the API for backward
+   * compatibility.
+   *
+   * @param jobDeployment the deployment to patch with
+   */
+  private void setDeploymentEnabled(JobDeployment jobDeployment) {
+    jobDeployment.setEnabled(jobDeployment.getEnabled() == null || jobDeployment.getEnabled());
   }
 }
