+++ projects/control-service/projects/pipelines_control_service/src/test/java/com/vmware/taurus/service/deploy/DataJobDeploymentControllerTest.java
@@ -10,11 +10,10 @@ import com.vmware.taurus.datajobs.DataJobsDeploymentController;
 import com.vmware.taurus.datajobs.TestUtils;
 import com.vmware.taurus.datajobs.ToModelApiConverter;
 import com.vmware.taurus.exception.DataJobDeploymentNotFoundException;
-import com.vmware.taurus.service.model.ActualDataJobDeployment;
 import com.vmware.taurus.service.model.DataJob;
 import com.vmware.taurus.service.model.DataJobDeploymentResources;
+import com.vmware.taurus.service.model.DesiredDataJobDeployment;
 import com.vmware.taurus.service.model.JobConfig;
-import com.vmware.taurus.service.repository.ActualJobDeploymentRepository;
 import com.vmware.taurus.service.repository.DesiredJobDeploymentRepository;
 import com.vmware.taurus.service.repository.JobsRepository;
 import org.junit.jupiter.api.AfterEach;
@@ -31,7 +30,6 @@ public class DataJobDeploymentControllerTest {
   @Autowired JobsRepository jobsRepository;
   @Autowired DataJobsDeploymentController dataJobsDeploymentController;
   @Autowired DesiredJobDeploymentRepository desiredJobDeploymentRepository;
-  @Autowired ActualJobDeploymentRepository actualJobDeploymentRepository;
 
   @MockBean DeploymentService deploymentService;
 
@@ -54,8 +52,8 @@ public class DataJobDeploymentControllerTest {
   public void testDeploymentPatch_expectDeploymentMerge() {
     var dataJob = ToModelApiConverter.toDataJob(TestUtils.getDataJob("teamName", "jobName"));
     jobsRepository.save(dataJob);
-    var existingDeployment = getActualJobDeployment(dataJob);
-    actualJobDeploymentRepository.save(existingDeployment);
+    var existingDeployment = getDesiredJobDeployment(dataJob);
+    desiredJobDeploymentRepository.save(existingDeployment);
 
     var newDeployment = TestUtils.getDataJobDeployment("deploymentId", "jobVersion");
     dataJobsDeploymentController.deploymentPatch("teamName", "jobName", "id", newDeployment);
@@ -82,8 +80,8 @@ public class DataJobDeploymentControllerTest {
     jobsRepository.deleteAll();
   }
 
-  private ActualDataJobDeployment getActualJobDeployment(DataJob dataJob) {
-    var deployment = new ActualDataJobDeployment();
+  private DesiredDataJobDeployment getDesiredJobDeployment(DataJob dataJob) {
+    var deployment = new DesiredDataJobDeployment();
     deployment.setGitCommitSha("actualSha");
     deployment.setDataJob(dataJob);
     deployment.setDataJobName(dataJob.getName());
